change functions from a file to directory
