+++ python/rikai/spark/functions/vision.py
@@ -1,4 +1,4 @@
-#  Copyright 2020 Rikai Authors
+#  Copyright 2021 Rikai Authors
 #
 #  Licensed under the Apache License, Version 2.0 (the "License");
 #  you may not use this file except in compliance with the License.
@@ -12,23 +12,24 @@
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
 
-"""Pyspark UDFs."""
+"""Vision-related Pyspark UDFs
+"""
 
+# Third Party
 from pyspark.sql.functions import udf
 from pyspark.sql.types import FloatType, StringType
 
 # Rikai
-from rikai.io import copy as _copy
 from rikai.logging import logger
-from rikai.spark.types import ImageType, LabelType
-from rikai.types import Box2d, Image, Label
-
-__all__ = ["label", "area", "copy", "image_copy"]
+from rikai.io import copy as _copy
+from rikai.spark.types.vision import ImageType, LabelType
+from rikai.types.geometry import Box2d
+from rikai.types.vision import Image, Label
 
 
 @udf(returnType=LabelType())
 def label(value: str) -> Label:
-    """Convert a string value into label"""
+    """Convert a string value into :py:class:`Label`."""
     return Label(value)
 
 
@@ -36,26 +37,6 @@ def label(value: str) -> Label:
 def area(bbox: Box2d) -> float:
     """A UDF to calculate the area of a bounding box"""
     return bbox.area
-
-
-@udf(returnType=StringType())
-def copy(source: str, dest: str) -> str:
-    """Copy a file from source to dest
-
-    Parameters
-    ----------
-    source : str
-        The source URI to copy from
-    dest : str
-        The destination uri or the destionation directory. If ``dest`` is
-        a URI ends with a "/", it represents a directory.
-
-    Return
-    ------
-    str
-        Return the URI of destination.
-    """
-    return _copy(source, dest)
 
 
 @udf(returnType=ImageType())
